chg: dev: Add tap interfaces to container platform
This change adds interfaces to the container platform in swns namespace.
The code checks if the interface exists and if it doesn't, then adds a tap interface.
This allows interfaces on the container to be shut/no shut, assign IP addresses and
self ping. Potentially, this change can help developers to pump some traffic from their VMs into the switch instance.

Change-Id: Idaf1a5931873342e2d34b7c846c96e49379fc296
Signed-off-by: Nachiket Abhyankar <nabhyank@hpe.com>