Eliminated unspported vlan items in mirror_set
Segfaults have been seen in switchd due to the container
plugin attempting to clone a src_vlans bitmap coming from a
variable that is not initialized before calling mirror_set.
This eliminates the vlan-related code in the mirror_set API
to avoid the crash (prevents Taiga 948 from occurring). A
separate change in switchd will address the uninitialized
variable.

Change-Id: I68820ac30db14ea78ce07b1724e0e73a9a170bac
Signed-off-by: jeff.mcmenomey <jeff.mcmenomey@hpe.com>